+++ src/EvergridLayout.ts
@@ -11,6 +11,7 @@ import {
     PanResponderInstance,
 } from 'react-native';
 import { WeakRef } from '@ungap/weakrefs';
+import merge from 'lodash.merge';
 import {
     Evergrid,
     ItemView,
@@ -28,6 +29,7 @@ import {
     PanPressableCallbacks,
     IInsets,
     LayoutLinkAxis,
+    LayoutLinkOptions,
 } from './types';
 import { insetPoint, insetSize, insetTranslation, zeroPoint } from './util';
 import {
@@ -160,6 +162,10 @@ export interface EvergridLayoutProps extends EvergridLayoutPrimitiveProps {
      **/
     anchor?: AnimatedValueXYDerivedInput<EvergridLayout>;
     /**
+     * Allows synrchonizing grids along the x, y or both axes.
+     */
+    linkedLayouts?: (LayoutLinkOptions & { layout: EvergridLayout })[];
+    /**
      * Modify the pan target.
      * Defaults to [viewOffset]{@link EvergridLayout#viewOffset}
      */
@@ -232,7 +238,7 @@ export default class EvergridLayout {
 
     private _weakViewRef?: WeakRef<Evergrid>;
 
-    private _locationOffsetBase$: Animated.ValueXY;
+    readonly locationOffsetBase$: Animated.ValueXY;
     private _locationOffsetBase: IPoint;
     private _scale: IPoint;
     private _hasScale = false;
@@ -272,6 +278,11 @@ export default class EvergridLayout {
     } = { x: [], y: [], xy: [] };
 
     constructor(options?: EvergridLayoutCallbacks & EvergridLayoutProps) {
+        let linkProps: Partial<EvergridLayoutProps> = {};
+        if (options?.linkedLayouts) {
+            // Extract linked properties
+            linkProps = this._getLinkLayoutProps(options.linkedLayouts);
+        }
         let {
             layoutSources,
             offset,
@@ -290,7 +301,7 @@ export default class EvergridLayout {
             snapToLocation,
             onViewportSizeChanged,
             onScaleChanged,
-        } = { ...kDefaultProps, ...options };
+        } = { ...kDefaultProps, ...linkProps, ...options };
 
         this._layoutSources = [];
 
@@ -399,23 +410,23 @@ export default class EvergridLayout {
         });
         this._animatedSubscriptions[sub] = this.anchor$;
 
-        this._locationOffsetBase$ = normalizeAnimatedDerivedValueXY(offset, {
+        this.locationOffsetBase$ = normalizeAnimatedDerivedValueXY(offset, {
             info: this,
         });
         this._locationOffsetBase = {
             // @ts-ignore: _value is private
-            x: this._locationOffsetBase$.x._value || 0,
+            x: this.locationOffsetBase$.x._value || 0,
             // @ts-ignore: _value is private
-            y: this._locationOffsetBase$.y._value || 0,
+            y: this.locationOffsetBase$.y._value || 0,
         };
         let _previousLocation = { ...this._locationOffsetBase };
-        sub = this._locationOffsetBase$.addListener(p => {
+        sub = this.locationOffsetBase$.addListener(p => {
             _previousLocation.x = this._locationOffsetBase.x;
             _previousLocation.y = this._locationOffsetBase.y;
             this._locationOffsetBase = p;
             this._didChangeLocationOffsetBase(_previousLocation, p);
         });
-        this._animatedSubscriptions[sub] = this._locationOffsetBase$;
+        this._animatedSubscriptions[sub] = this.locationOffsetBase$;
 
         this._locationOffsetTarget = {};
 
@@ -504,6 +515,13 @@ export default class EvergridLayout {
         this._transferViewOffsetToLocation();
 
         this.setLayoutSources(layoutSources || []);
+
+        if (options?.linkedLayouts) {
+            // Complete layout links
+            for (let linkedLayout of options.linkedLayouts) {
+                this.linkLayout(linkedLayout);
+            }
+        }
     }
 
     /**
@@ -561,7 +579,8 @@ export default class EvergridLayout {
         return layouts;
     }
 
-    linkLayout(layout: EvergridLayout, options: { axis: LayoutLinkAxis }) {
+    linkLayout(options: LayoutLinkOptions & { layout: EvergridLayout }) {
+        const { layout } = options;
         if (!(layout instanceof EvergridLayout) || layout === this) {
             throw new Error('Invalid EvergridLayout');
         }
@@ -574,15 +593,96 @@ export default class EvergridLayout {
 
             this._checkLayoutLink(layout, options);
 
-            let cbOptions = { ...options, receiver: this };
-            this.didLinkWithLayout(layout, cbOptions);
-            layout.didLinkWithLayout(this, cbOptions);
+            this.didLinkWithLayout({ ...options, layout, receiver: this });
+            layout.didLinkWithLayout({
+                ...options,
+                layout: this,
+                receiver: this,
+            });
+
+            // Propagate links
+            for (let otherLayout of this.linkedLayouts(options.axis)) {
+                if (otherLayout === layout) {
+                    continue;
+                }
+                otherLayout.linkLayout(options);
+            }
+            for (let otherLayout of layout.linkedLayouts(options.axis)) {
+                if (otherLayout === layout) {
+                    continue;
+                }
+                otherLayout.linkLayout(options);
+            }
         }
+    }
+
+    /**
+     * Returns properties which should be used during
+     * layout intialization.
+     *
+     * @param layout
+     * @param options
+     * @returns
+     */
+    private _getLinkLayoutProps(
+        options: (LayoutLinkOptions & { layout: EvergridLayout })[]
+    ): Partial<EvergridLayoutProps> {
+        return options.reduce<Partial<EvergridLayoutProps>>(
+            (mergedOptions, options) =>
+                merge(mergedOptions, this._getLinkLayoutPropsSingle(options)),
+            {}
+        );
+    }
+
+    /**
+     * Returns properties which should be used during
+     * layout intialization.
+     *
+     * @param layout
+     * @param options
+     * @returns
+     */
+    private _getLinkLayoutPropsSingle(
+        options: LayoutLinkOptions & { layout: EvergridLayout }
+    ): Partial<EvergridLayoutProps> {
+        let axes: ('x' | 'y')[] = [];
+        if (options.axis === 'xy') {
+            axes = ['x', 'y'];
+        } else {
+            axes = [options.axis];
+        }
+        let props: Partial<EvergridLayoutProps> = {};
+        switch (options.mode || 'auto') {
+            case 'manual':
+                return props;
+            case 'auto':
+                for (let axis of axes) {
+                    // offset (locationOffsetBase)
+                    props.offset = {
+                        ...props.offset,
+                        [axis]: options.layout.locationOffsetBase$[axis],
+                    };
+
+                    // viewOffset
+                    props.viewOffset = {
+                        ...props.viewOffset,
+                        [axis]: options.layout.viewOffset$[axis],
+                    };
+
+                    // scale
+                    props.scale = {
+                        ...props.scale,
+                        [axis]: options.layout.scale$[axis],
+                    };
+                }
+                return props;
+        }
+        throw new Error('Invalid layout link mode');
     }
 
     private _checkLayoutLink(
         layout: EvergridLayout,
-        options: { axis: LayoutLinkAxis }
+        options: LayoutLinkOptions
     ) {
         let checkAxes: ('x' | 'y')[] = [];
         if (options.axis === 'xy') {
@@ -592,29 +692,31 @@ export default class EvergridLayout {
         }
         for (let axis of checkAxes) {
             if (
-                this._locationOffsetBase$[axis] !==
-                layout._locationOffsetBase$[axis]
+                this.locationOffsetBase$[axis] !==
+                layout.locationOffsetBase$[axis]
             ) {
                 console.warn(
-                    `You linked grids in the ${axis}-direction, but didn't use the same offset.${axis}. Create an animated value and use it in both grids.`
+                    `You linked grids in the ${axis}-direction, but didn't use the same offset.${axis}. Create an animated value and use it in both grids, or use linkedLayouts property for automatic setup.`
                 );
             }
             if (this.viewOffset$[axis] !== layout.viewOffset$[axis]) {
                 console.warn(
-                    `You linked grids in the ${axis}-direction, but didn't use the same viewOffset.${axis}. Create an animated value and use it in both grids.`
+                    `You linked grids in the ${axis}-direction, but didn't use the same viewOffset.${axis}. Create an animated value and use it in both grids, or use linkedLayouts property for automatic setup.`
                 );
             }
             if (this.scale$[axis] !== layout.scale$[axis]) {
                 console.warn(
-                    `You linked grids in the ${axis}-direction, but didn't use the same scale.${axis}. Create an animated value and use it in both grids.`
+                    `You linked grids in the ${axis}-direction, but didn't use the same scale.${axis}. Create an animated value and use it in both grids, or use linkedLayouts property for automatic setup.`
                 );
             }
         }
     }
 
     didLinkWithLayout(
-        layout: EvergridLayout,
-        options: { axis: LayoutLinkAxis; receiver: EvergridLayout }
+        options: LayoutLinkOptions & {
+            layout: EvergridLayout;
+            receiver: EvergridLayout;
+        }
     ) {}
 
     get view(): Evergrid {
@@ -935,7 +1037,7 @@ export default class EvergridLayout {
         };
         this._viewOffset = zeroPoint();
         this._locationOffsetBase = location;
-        this._locationOffsetBase$.setValue(location);
+        this.locationOffsetBase$.setValue(location);
         this.endUpdate();
     }
 
@@ -998,11 +1100,11 @@ export default class EvergridLayout {
     get locationOffset$(): IAnimatedPoint {
         return {
             x: Animated.add(
-                this._locationOffsetBase$.x,
+                this.locationOffsetBase$.x,
                 Animated.divide(this.viewOffset$.x, this.scale$.x)
             ),
             y: Animated.add(
-                this._locationOffsetBase$.y,
+                this.locationOffsetBase$.y,
                 Animated.divide(this.viewOffset$.y, this.scale$.y)
             ),
         };
@@ -1492,7 +1594,7 @@ export default class EvergridLayout {
             scale?: Partial<IAnimatedPoint> | Partial<Animated.ValueXY>;
         }
     ): IAnimatedPoint {
-        let { x: xl0, y: yl0 } = this._locationOffsetBase$;
+        let { x: xl0, y: yl0 } = this.locationOffsetBase$;
         let { x: x0, y: y0 } = this.containerOriginOffset$;
         let cp: IAnimatedPoint = {
             x: Animated.subtract(this.viewOffset$.x, x0),
@@ -1674,7 +1776,7 @@ export default class EvergridLayout {
 
         if (!options.animated) {
             this.beginUpdate();
-            needOffset && this._locationOffsetBase$.setValue(finalOffset);
+            needOffset && this.locationOffsetBase$.setValue(finalOffset);
             needScale && this.scale$.setValue(finalScale);
             let info = { finished: true };
             this._onEndDeceleration(info);
@@ -1687,7 +1789,7 @@ export default class EvergridLayout {
         let offsetAnimation: Animated.CompositeAnimation | undefined;
         if (needOffset) {
             if (options.timing) {
-                offsetAnimation = Animated.timing(this._locationOffsetBase$, {
+                offsetAnimation = Animated.timing(this.locationOffsetBase$, {
                     toValue: finalOffset,
                     easing: Easing.inOut(Easing.exp),
                     ...options.timing,
@@ -1695,7 +1797,7 @@ export default class EvergridLayout {
                 });
             } else {
                 offsetAnimation = Animated.spring(
-                    this._locationOffsetBase$, // Auto-multiplexed
+                    this.locationOffsetBase$, // Auto-multiplexed
                     {
                         toValue: finalOffset,
                         velocity: